More RNN blog work.
