+++ checkpoint
@@ -1,0 +1,2 @@
+model_checkpoint_path: "rnn.tflearn"
+all_model_checkpoint_paths: "rnn.tflearn"
